+++ rand-core/README.md
@@ -1,0 +1,21 @@
+rand-core
+====
+
+Core functionality for the [rand] library.
+
+This crate contains the core trait, `RngCore`, and some extension traits. This
+should be sufficient for libraries publishing an `RngCore` implementation,
+but most users should prefer to use the [rand] crate.
+Most parts of this crate are re-exported in [rand].
+
+[Documentation](https://docs.rs/rand-core)
+
+[rand]: ..
+
+
+# License
+
+`rand` is primarily distributed under the terms of both the MIT
+license and the Apache License (Version 2.0).
+
+See LICENSE-APACHE, and LICENSE-MIT for details.
