+++ src/prng/hc128.rs
@@ -11,8 +11,7 @@
 //! The HC-128 random number generator.
 
 use core::fmt;
-use {RngCore, CryptoRng, SeedableRng};
-use {impls, le};
+use rand_core::{RngCore, CryptoRng, SeedableRng, impls, le};
 
 const SEED_WORDS: usize = 8; // 128 bit key followed by 128 bit iv
